Float -> Double

+++ VeDe-Graphics/custompainter.h
@@ -11,22 +11,22 @@ namespace gx
 class CustomPainter
 {
 public:
-    virtual void drawLine(float x1, float y1, float x2, float y2) = 0;
+    virtual void drawLine(double x1, double y1, double x2, double y2) = 0;
     void drawLine(Vertex start, Vertex end);
 
-    virtual void drawEllipse(float cX, float cY, float rX, float rY) = 0;
+    virtual void drawEllipse(double cX, double cY, double rX, double rY) = 0;
     void drawEllipse(Vertex center, Vertex radius);
 
-    virtual void drawRectangle(float upLeftX, float upLeftY, float downRightX, float downRightY) = 0;
+    virtual void drawRectangle(double upLeftX, double upLeftY, double downRightX, double downRightY) = 0;
     void drawRectangle(Vertex upLeft, Vertex downRight);
 
     virtual void drawPath(QList<Vertex> vertices, QList<bool> moves) = 0;
     virtual void drawPath(QPainterPath& path) = 0;
 
-    virtual void setStrokeWidth(float width) = 0;
+    virtual void setStrokeWidth(double width) = 0;
     virtual void setStrokeColor(const Color& color) = 0;
     virtual void setBackColor(const Color& color) = 0;
-    virtual void setZoomFactor(float zoomFactor) = 0;
+    virtual void setZoomFactor(double zoomFactor) = 0;
 };
 }
