+++ VeDe-Graphics/objects/shape.h
@@ -16,12 +16,12 @@ public:
     void setGuiElement(bool guiElement);
 
 protected:
-    void fixBoxForStrokeWidth(QRectF& rect, float factor=0.5f) const;
+    void fixBoxForStrokeWidth(QRectF& rect, double factor=0.5f) const;
 private:
     bool m_guiElement;
     uint m_canvasZoomHandlerId;
 
-    void onCanvasZoomChange(float zoomFactor);
+    void onCanvasZoomChange(double zoomFactor);
 };
 }
