+++ VeDe-Graphics/vertex.cpp
@@ -12,27 +12,27 @@ gx::Vertex::Vertex(const Vertex &other)
 {
 }
 
-gx::Vertex::Vertex(float x, float y)
+gx::Vertex::Vertex(double x, double y)
     :m_x(x), m_y(y)
 {
 }
 
-float gx::Vertex::x() const
+double gx::Vertex::x() const
 {
     return m_x;
 }
 
-void gx::Vertex::setX(float x)
+void gx::Vertex::setX(double x)
 {
     m_x = x;
 }
 
-float gx::Vertex::y() const
+double gx::Vertex::y() const
 {
     return m_y;
 }
 
-void gx::Vertex::setY(float y)
+void gx::Vertex::setY(double y)
 {
     m_y = y;
 }
@@ -61,70 +61,70 @@ gx::Vertex gx::Vertex::operator -(const gx::Vertex &other) const
     return Vertex(m_x - other.x(), m_y - other.y());
 }
 
-float gx::Vertex::operator *(const gx::Vertex &other) const
+double gx::Vertex::operator *(const gx::Vertex &other) const
 {
     return m_x * other.x() + m_y * other.y();
 }
 
-gx::Vertex &gx::Vertex::operator +=(float &scalar)
+gx::Vertex &gx::Vertex::operator +=(double &scalar)
 {
     this->m_x += scalar;
     this->m_y += scalar;
     return *this;
 }
 
-gx::Vertex &gx::Vertex::operator -=(float &scalar)
+gx::Vertex &gx::Vertex::operator -=(double &scalar)
 {
     this->m_x -= scalar;
     this->m_y -= scalar;
     return *this;
 }
 
-gx::Vertex &gx::Vertex::operator *=(float &scalar)
+gx::Vertex &gx::Vertex::operator *=(double &scalar)
 {
     this->m_x *= scalar;
     this->m_y *= scalar;
     return *this;
 }
 
-gx::Vertex &gx::Vertex::operator /=(float &scalar)
+gx::Vertex &gx::Vertex::operator /=(double &scalar)
 {
     this->m_x /= scalar;
     this->m_y /= scalar;
     return *this;
 }
 
-gx::Vertex gx::Vertex::operator +(float &scalar) const
+gx::Vertex gx::Vertex::operator +(double &scalar) const
 {
     return Vertex(m_x + scalar, m_y + scalar);
 }
 
-gx::Vertex gx::Vertex::operator -(float &scalar) const
+gx::Vertex gx::Vertex::operator -(double &scalar) const
 {
     return Vertex(m_x - scalar, m_y - scalar);
 }
 
-gx::Vertex gx::Vertex::operator *(float &scalar) const
+gx::Vertex gx::Vertex::operator *(double &scalar) const
 {
     return Vertex(m_x * scalar, m_y * scalar);
 }
 
-gx::Vertex gx::Vertex::operator /(float &scalar) const
+gx::Vertex gx::Vertex::operator /(double &scalar) const
 {
     return Vertex(m_x / scalar, m_y / scalar);
 }
 
-float gx::Vertex::distance(const gx::Vertex &other) const
+double gx::Vertex::distance(const gx::Vertex &other) const
 {
     return qSqrt(qPow(other.x() - m_x, 2) + qPow(other.y() - m_y, 2));
 }
 
-float gx::Vertex::length() const
+double gx::Vertex::length() const
 {
     return qSqrt(qPow(m_x, 2) + qPow(m_y, 2));
 }
 
-float gx::Vertex::lengthSquared() const
+double gx::Vertex::lengthSquared() const
 {
     return qPow(m_x, 2) + qPow(m_y, 2);
 }
